+++ src/runtime/test/test_ConfigsMerger.cpp
@@ -368,14 +368,16 @@ TEST(ConfigsMergerTestGroup, bundle_annotations) {
         CHECK((ConfigsMerger{config, metadata}.getBundleAnnotations() == expectedAnnotations));
         metadata.labels.erase("com.test.image.key");
     }
-    // Engine annotations override image ones
+    // Custom annotations (from the CLI or from components like the FileDescriptoHandler)
+    // override everything
     {
         metadata.labels["com.test.dummy_key"] = "image_value";
         auto configRAII = test_utility::config::makeConfig();
         auto& config = configRAII.config;
+        config->commandRun.ociAnnotations["com.hooks.logging.level"] = "0";
         auto expectedAnnotations = std::unordered_map<std::string, std::string>{
             {"com.test.dummy_key", "dummy_value"},
-            {"com.hooks.logging.level", "2"}
+            {"com.hooks.logging.level", "0"}
         };
         auto annot = ConfigsMerger{config, metadata}.getBundleAnnotations();
         CHECK((ConfigsMerger{config, metadata}.getBundleAnnotations() == expectedAnnotations));